// src/app/trainee-biodata/biodata-tab-layout/models/GrandFather.ts
export interface GrandFather {
    grandFatherId: number;
    traineeId: number;  
    grandFatherTypeId: number; 
    grandFatherType: string; 
    occupation: string; 
    deadStatusValue: string; 
    nationality: string;  
    grandFathersName: string;    
    occupationId: number;
    age: number;
    deadStatus: number;
    nationalityId: number;
    additionalInformation: string;    
    status: number;
    menuPosition: number;
    isActive: boolean;
} 